feat: solve planting and unplanting tree in logger
Tree unplanting and planting once again is necessary, since otherwise old planted trees in scopes above (without the correct authenticator) log and try to rollover the log files.
A necessary pain of this is to have get_it as a dependency in logger package :/ This makes the tree handling much simpler though. Can maybe be removed once we migrate from Fimber.