chore(mock): 2266 is adapted for lightweight user testings (workaround - to be reverted)
Due to bug in DAS client no ADL over multiple SP can be parsed in 0.24.1, hence I joined the necessary segments together for a quick workaround and added events with JP updates for better UX testing experience.

This will all need to be reverted once the bugs in ADL are fixed.